more client i2c shuffling
