+++ client/cmd.cpp
@@ -4,10 +4,9 @@
 #include <string.h>
 
 #include "cmd.h"
-#include "i2ctcpv1.h"
 #include "pb/types.h"
 #include "rl.h"
-#include "sock.h"
+#include "i2c.h"
 #include "parse.h"
 
 #include "pb/bus.h"
