+++ client/sock.cpp
@@ -7,13 +7,12 @@
 #include <sys/socket.h>
 #include <netinet/in.h>
 #include <errno.h>
-
 #include <thread>
 
 #include "i2ctcpv1.h"
 #include "sock.h"
 #include "rl.h"
-#include "xxd.h"
+#include "i2c.h"
 
 using std::logic_error;
 using std::thread;
@@ -104,33 +103,5 @@ void PBSocket::sock_task() {
 	}
 
 	sock_close();
-}
-
-bool i2c_dump_send = false;
-bool i2c_dump_recv = true;
-
-void i2c_send(uint16_t addr, const char * data, size_t data_size) {
-	i2ctcp_msg_t msg = {
-		.addr = addr,
-		.data = (char *) data,
-		.length = data_size,
-	};
-
-	char* packed;
-	size_t size;
-	if (!i2ctcp_write(&msg, &packed, &size)) return;
-
-	sock->send(packed, size);
-	if (i2c_dump_send) {
-		printf("[i2c send] data(0x%02lx):\n", data_size);
-		xxd(data, data_size);
-	}
-}
-
-void i2c_recv(uint16_t addr, const char * data, size_t data_size) {
-	if (i2c_dump_recv) {
-		printf("[i2c recv] data(0x%02lx):\n", data_size);
-		xxd(data, data_size);
-	}
 }
