+++ client/sock.h
@@ -29,6 +29,3 @@ private:
 
 extern PBSocket* sock;
 
-void i2c_send(uint16_t addr, const char * data, size_t data_size);
-void i2c_recv(uint16_t addr, const char * data, size_t data_size);
-
