+++ shared/include.cmake
@@ -1,5 +1,5 @@
 include_directories(${CMAKE_CURRENT_LIST_DIR})
 add_library(puzbus STATIC
-	${CMAKE_CURRENT_LIST_DIR}/pb/driver.c
+	${CMAKE_CURRENT_LIST_DIR}/pb/moddrv.c
 	)
